Threading now works!
The three threads are now communicating in a healthy manner. Fades work properly and now overwrite one another like they should

Up next:
When an index changes value, it should make sure that any queue items that remain after it is altered get removed. Right now, if a long fade is taking place and a short fade overwrites it, at the conclusion of the short fade, it will pop back into the long fade.