Fetchloop and Clockloop Tweaks
Timing on ClockLoop now uses epoch time instead of sleep(). This should deliver more precise timing on the framerate.

Fetchloop has an increased sock.listen parameter, and added a task_done on the command queue operation. Hopefully this will increase the number of commands that the server is able to handle every second.